+++ app/src/main/java/com/example/randomuser/components/DetailsView.kt
@@ -7,14 +7,13 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBackIosNew
 import androidx.compose.runtime.*
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.randomuser.data.UserItemData
+import com.example.randomuser.exampledata.exampleResults
 import com.example.randomuser.helpers.addFirstAndLastName
 import com.example.randomuser.model.Results
 import com.example.randomuser.ui.theme.randomuserTheme
@@ -51,7 +50,6 @@ fun DetailsView(
     }
 }
 
-/*
 @SuppressLint("UnrememberedMutableState")
 @Preview("ComponentPreview (light)", showBackground = true)
 @Preview("ComponentPreview (dark)", uiMode = Configuration.UI_MODE_NIGHT_YES)
@@ -59,9 +57,7 @@ fun DetailsView(
 @Preview("ComponentPreview (large screen)", device = Devices.PIXEL_C)
 @Composable
 fun DetailsPreview() {
-    val currentUser = rememberSaveable { mutableStateOf(UserItemData("Pelle", "Anderson", "hello@hi.com", "BR", "https://randomuser.me/api/portraits/men/57.jpg", 54, "male", 4)) }
     randomuserTheme {
-        DetailsView(onClick = {}, currentUser = mutableStateOf(currentUser.value!!))
+        DetailsView(onClick = {}, currentUser = mutableStateOf(exampleResults))
     }
 }
-*/
